+++ google/cloud/datastore/__init__.py
@@ -61,6 +61,16 @@ from google.cloud.datastore.client import Client
 from google.cloud.datastore.entity import Entity
 from google.cloud.datastore.key import Key
 from google.cloud.datastore.query import Query
+from google.cloud.datastore.query_profile import ExplainOptions
 from google.cloud.datastore.transaction import Transaction
 
-__all__ = ["__version__", "Batch", "Client", "Entity", "Key", "Query", "Transaction"]
+__all__ = [
+    "__version__",
+    "Batch",
+    "Client",
+    "Entity",
+    "Key",
+    "Query",
+    "ExplainOptions",
+    "Transaction",
+]
