+++ tests/unit/test_query.py
@@ -1079,7 +1079,7 @@ def test_iterator_sends_explain_options_w_request(database_id, analyze):
     When query has explain_options set, all requests should include
     the explain_options field.
     """
-    from google.cloud.datastore.query import ExplainOptions
+    from google.cloud.datastore.query_profile import ExplainOptions
 
     response_pb = _make_query_response([], b"", 0, 0)
     ds_api = _make_datastore_api(response_pb)
@@ -1100,7 +1100,7 @@ def test_iterator_explain_metrics(database_id):
     """
     If explain_metrics is recieved from backend, it should be set on the iterator
     """
-    from google.cloud.datastore.query import ExplainMetrics
+    from google.cloud.datastore.query_profile import ExplainMetrics
     from google.cloud.datastore_v1.types import query_profile as query_profile_pb2
     from google.protobuf import duration_pb2
 
@@ -1132,7 +1132,7 @@ def test_iterator_explain_metrics_no_explain(database_id):
     If query has no explain_options set, iterator.explain_metrics should raise
     an exception.
     """
-    from google.cloud.datastore.query import QueryExplainError
+    from google.cloud.datastore.query_profile import QueryExplainError
 
     ds_api = _make_datastore_api()
     client = _Client(None, datastore_api=ds_api)
@@ -1153,8 +1153,8 @@ def test_iterator_explain_metrics_no_analyze_make_call(database_id):
     If query.explain_options(analyze=False), accessing iterator.explain_metrics
     should make a network call to get the data.
     """
-    from google.cloud.datastore.query import ExplainOptions
-    from google.cloud.datastore.query import ExplainMetrics
+    from google.cloud.datastore.query_profile import ExplainOptions
+    from google.cloud.datastore.query_profile import ExplainMetrics
     from google.cloud.datastore_v1.types import query_profile as query_profile_pb2
     from google.protobuf import duration_pb2
 
@@ -1503,7 +1503,7 @@ def _make_datastore_api(*results):
 
 
 def test_explain_metrics__from_pb():
-    from google.cloud.datastore.query import (
+    from google.cloud.datastore.query_profile import (
         ExplainMetrics,
         _ExplainAnalyzeMetrics,
         QueryExplainError,
@@ -1550,7 +1550,7 @@ def test_explain_metrics_execution_stats():
     Standard ExplainMetrics class should raise exception when execution_stats is accessed.
     _ExplainAnalyzeMetrics should include the field
     """
-    from google.cloud.datastore.query import (
+    from google.cloud.datastore.query_profile import (
         ExplainMetrics,
         QueryExplainError,
         _ExplainAnalyzeMetrics,
@@ -1573,7 +1573,7 @@ def test_explain_options__to_dict():
     """
     Should be able to create a dict representation of ExplainOptions
     """
-    from google.cloud.datastore.query import ExplainOptions
+    from google.cloud.datastore.query_profile import ExplainOptions
 
     assert ExplainOptions(analyze=True)._to_dict() == {"analyze": True}
     assert ExplainOptions(analyze=False)._to_dict() == {"analyze": False}
